+++ auth.go
@@ -7,8 +7,12 @@ import (
 	"go.uber.org/zap"
 )
 
-// ErrNotAllowed is returned if the user is not allowed access to a specific scope.
-const ErrNotAllowed = Error("not allowed")
+const (
+	// ErrNotAllowed is returned if the user is not allowed access to a specific scope.
+	ErrNotAllowed = Error("not allowed")
+
+	permissionKeyPrefix = "joe.permissions."
+)
 
 // Auth implements logic to add user authorization checks to your bot.
 type Auth struct {
@@ -60,6 +64,41 @@ func (a *Auth) CheckPermission(scope, userID string) error {
 	}
 
 	return ErrNotAllowed
+}
+
+// Users returns a list of userIDs having one or more permission scopes
+func (a *Auth) Users() ([]string, error) {
+	a.logger.Debug("Retrieving all userIDs")
+
+	keys, err := a.store.Keys()
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to load permissions")
+	}
+
+	var userIDs []string
+	for _, key := range keys {
+		if strings.HasPrefix(key, permissionKeyPrefix) {
+			userID := strings.TrimPrefix(key, permissionKeyPrefix)
+			userIDs = append(userIDs, userID)
+		}
+	}
+
+	return userIDs, nil
+}
+
+// UserPermissions returns the permission scopes for a specific user
+func (a *Auth) UserPermissions(userID string) ([]string, error) {
+	a.logger.Debug("Retrieving user permissions",
+		zap.String("user_id", userID),
+	)
+
+	key := a.permissionsKey(userID)
+	permissions, err := a.loadPermissions(key)
+	if err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	return permissions, nil
 }
 
 func (a *Auth) loadPermissions(key string) ([]string, error) {
@@ -188,5 +227,5 @@ func (a *Auth) updatePermissions(key string, permissions []string) error {
 }
 
 func (a *Auth) permissionsKey(userID string) string {
-	return "joe.permissions." + userID
+	return permissionKeyPrefix + userID
 }